+++ src/findmy_rest/__init__.py
@@ -1,3 +1,3 @@
 """REST API for Apple Find My accessory and device location and battery."""
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"